# back-end/tel_openapi.yaml
openapi: "3.0.2"
info:
  description: Control Center API description
  title: Control Center
  version: 1.0.0
paths:
...
  /incidents:
    get:
      tags:
        - incidents
      summary: Get all incidents available
      operationId: get_incidents
      responses:
        '200':
          description: Successful retrieval of incidents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/incidentArray"
        '404':
          description: Not Found
        '401':
          description: Not Authorized for this operation
        '500':
          description: Internal server error
    post:
      tags:
        - incidents
      summary: Submit a new incident
      description: When an incident is submitted, it is considered open and user are deployed according to the agencies and the levels requested.
      operationId: post_incidents
      requestBody:
        content:
          application/json:    
            schema:
              $ref: "#/components/schemas/incident"
        required: true
      responses:
        '200':
          description: Successful submition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/incident"
        '500':
          description: Internal server error
        '401':
          description: Not Authorized for this operation
        '400':
          description: Bad request
        '403':
          description: Forbidden to use this endpoint
  /incidents/{incidentId}:
    get:
      tags:
        - incidentsId
      summary: Get an incident
      operationId: get_incidents_byid
      parameters:
        - name: incidentId
          in: path
          description: ID of incident
          required: true
          schema:
            $ref: "#/components/schemas/incidentId"
      responses:
        '200':
          description: Successful retrieval of wanted incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/incident"
        '404':
          description: Incident not found
        '401':
          description: Not Authorized for this operation
    delete:
      tags:
        - incidentsId
      summary: Delete an incident
      operationId: delete_incidents_byid
      parameters:
        - name: incidentId
          in: path
          description: ID of incident
          required: true
          schema:
            $ref: "#/components/schemas/incidentId"
      responses:
        '200':
          description: Successful deletion of wanted incident
        '404':
          description: Incident not found
        '401':
          description: Not Authorized for this operation
        '403':
          description: Forbidden to use this endpoint
    put:
      tags:
        - incidentsId
      summary: Update an incident
      operationId: put_incidents_byid
      parameters:
        - name: incidentId
          in: path
          description: ID of incident
          required: true
          schema:
            $ref: "#/components/schemas/incidentId"
      requestBody:
        content:
          application/json:    
            schema:
              $ref: "#/components/schemas/incident"
      responses:
        '200':
          description: Successful updated incident
        '404':
          description: Incident not found
        '401':
          description: Not Authorized for this operation
        '403':
          description: Forbidden to use this endpoint
    patch:
      tags:
        - incidentsId
      summary: Update an incident
      description: This endpoint is used to add more available users of any agency to the incident.
      operationId: patch_incidents_byid
      parameters:
        - name: incidentId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/incidentId"
      requestBody:
        content:
          application/json:    
            schema:
              $ref: "#/components/schemas/addUsers"
      responses:
        '200':
          description: Successful update
        '404':
          description: Not Found
        '401':
          description: Not Authorized for this operation
        '403':
          description: Forbidden to use this endpoint
        '500':
          description: Internal server error
...
components:
  schemas:
...
    userId:
      type: string
      description: String containing the user id
    incidentId:
      type: string
      description: String containing the incident id
...
    userArrayPerAgency:
      type: object
      description: Object containing pairs of (agencyId - array of users), keeping all users currently in the incident per agency.
      example:
        5ee8ebb9dbbb2461557d2983: "array of users for this agency.."
        5ee8ebb9dbbb2461557d2984: "array of users for this agency.."
    incident:
      type: object
      required:
        - incidentId
      properties:
        incidentId:
          $ref: "#/components/schemas/incidentId"
        title:
          type: string
          description: String containing the unique incident title
        x:
          type: number
          description: Longtitude of the position of the incident
        y:
          type: number
          description: Latitude of the position of the incident
        startDate:
          type: string
        endDate:
          type: string
        telephone:
          type: string
          description: Telephone of the caller that initiated the incident
        level:
          $ref : "#/components/schemas/levels"
        isOpen:
          type: boolean
        userId:
          $ref: "#/components/schemas/userId"
        users:
          $ref: "#/components/schemas/userArrayPerAgency"
    incidentArray:
      type: array
      items:
        $ref: "#/components/schemas/incident"
...
    levels:
      type: object
      description: JSON object containing pairs of (agencyId - level given)
      example:
        5ee8ebb9dbbb2461557d2983: 1
        5ee8ebb9dbbb2461557d2984: 2
    addUsers:
      type: object
      description: JSON object containing (agency - number of users to add) pairs
      example:
        addUsers:
          5ee8cee602d2ee3624835211: 5
          5ee8cee602d2ee3624835212: 7
...